diplom3

+++ sprint4/problems/leave_game/solution/sql/types.h
@@ -12,5 +12,5 @@ struct Record {
     UUIDType id;
     std::string name;
     int score;
-    size_t play_time;
+    double play_time;
 };
